% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.model.R
\name{citrus.buildFoldsEndpointModels}
\alias{citrus.buildFoldsEndpointModels}
\title{Build models from each fold of clustering}
\usage{
citrus.buildFoldsEndpointModels(
  type,
  citrus.foldFeatureSet,
  labels,
  regularizationThresholds = NULL,
  family = "classification",
  ...
)
}
\arguments{
\item{type}{Model Type. Valid options are \code{pamr}, \code{glmnet}, and \code{sam}.}

\item{labels}{Endpoint labels for samples.}

\item{regularizationThresholds}{Regularization thresholds for penalized models.}

\item{family}{Family of model to be constructed. Valid options are \code{classification} and \code{continuous}.}

\item{...}{Other arguments passed to model-fitting functions.}

\item{citrus.foldFeatureSet.}{A \code{citrus.foldFeatureSet} object.}
}
\value{
A list of models, one model fit on each fold's feature set.
}
\description{
Builds a model from features derived from each independent fold of clustering.
}
\author{
Robert Bruggner
}
