% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.model.R
\name{citrus.extractModelFeatures}
\alias{citrus.extractModelFeatures}
\title{Report model features at pre-specified thresholds.}
\usage{
citrus.extractModelFeatures(cvMinima, finalModel, finalFeatures)
}
\arguments{
\item{cvMinima}{List of regularization indicies at which to extract model features, produced by \code{\link{citrus.getCVMinima}}.}

\item{finalModel}{Predictive model from which to extract non-zero features.}

\item{finalFeatures}{Features used to construct \code{finalModel}.}
}
\value{
List of significant features and clusters at specified thresholds.
}
\description{
Report model features at pre-specific thresholds. For predictive models, reports non-zero model features at specified
regularization thresholds. For FDR-constrained models, reports features below specified false discovery rates.
}
\author{
Robert Bruggner
}
