% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.util.R
\name{citrus.familyList}
\alias{citrus.familyList}
\title{List possible model families}
\usage{
citrus.familyList()
}
\value{
Vector of model families that Citrus can compute.
}
\description{
Returns valid model family types that Citrus can compute.
}
\author{
Robert Bruggner
}

