% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.model.R
\name{citrus.getCVMinima}
\alias{citrus.getCVMinima}
\title{Get regularization thresholds of pre-selected cross-validation points}
\usage{
citrus.getCVMinima(modelType, thresholdCVRates, fdrRate = 0.01)
}
\arguments{
\item{modelType}{Method to be used for model-fitting. Valid options are: \code{glmnet},\code{pamr}, and \code{sam}.}

\item{thresholdCVRates}{Matrix of error rates at regularizationThresholds returned by \code{citrus.thresholdCVs.*} function.}

\item{fdrRate}{FDR Maximum used to determine FDR-constrained model regularization threshold.}
}
\value{
List of regularization thresholds and indicies based on pre-selected cross-validation error rates points.
}
\description{
#' Get regularization thresholds of pre-selected cross-validation points and their indicies.
}
\details{
For predictive models (i.e. \code{pamr} or \code{glmnet}), returns indicies of regularization thresholds
producing the minimum cross validation error rate (\code{cv.min}), the simplest model having error within 1
standard error of the minimum (\code{cv.1se}), and the model with the minimum error having an FDR rate < \code{fdrRate} (\code{cv.fdr.constrained})
when possible.
}
\author{
Robert Bruggner
}
