% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.util.R
\name{citrus.modelTypes}
\alias{citrus.modelTypes}
\title{List possible model types}
\usage{
citrus.modelTypes()
}
\value{
Vector of model types that Citrus can compute.
}
\description{
Returns valid model types that Citrus can compute.
}
\author{
Robert Bruggner
}
