% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.classModel.R,
%   R/citrus.continuousModel.R, R/citrus.model.R
\name{citrus.predict}
\alias{citrus.predict}
\alias{citrus.predict.classification}
\alias{citrus.predict.continuous}
\title{Predict labels of new feature set}
\usage{
citrus.predict.classification(citrus.endpointModel, newFeatures)

citrus.predict.continuous(citrus.endpointModel, newFeatures)

citrus.predict(citrus.endpointModel, newFeatures)
}
\arguments{
\item{citrus.endpointModel}{A \code{citrus.endpointModel} object.}

\item{newFeatures}{Features from samples to predict labels for.}
}
\value{
Matrix of predicted sample endpoints at all model regularization thresholds.
}
\description{
Predict labels of new feature set
}
\author{
Robert Bruggner
}
