% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.util.R
\name{citrus.selectClusters.minimumClusterSize}
\alias{citrus.selectClusters.minimumClusterSize}
\title{Selects clusters for endpoint analysis}
\usage{
citrus.selectClusters.minimumClusterSize(
  citrus.clustering,
  minimumClusterSizePercent = 0.05,
  ...
)
}
\arguments{
\item{citrus.clustering}{A \code{citrus.clustering} object.}

\item{minimumClusterSizePercent}{The percentage \eqn{0 < x < 1} of the total number of clustered events a cluster 
must contain in order to be selected.}

\item{...}{Other arguments (ignored).}
}
\description{
Selects clusters for endpoint analysis by cluster size. Selected clusters must have a minimum number of 
cells in them as a proportion of the total number of clustered events. If \eqn{n} total events are clustered,
clusters contatining at least \eqn{n * minimumClusterSizePercent} events are selected.
}
\author{
Robert Bruggner
}
