% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.util.R
\name{citrus.selectClusters}
\alias{citrus.selectClusters}
\title{Select clusters for further analysis}
\usage{
citrus.selectClusters(citrus.clustering, method = "minimumClusterSize", ...)
}
\arguments{
\item{citrus.clustering}{A \code{citrus.clustering} object.}

\item{method}{Method for determining which clusters from a clustering should be analyzed for stratifying signal.}

\item{...}{Other parameters passed to specific cluster selection methods.}
}
\value{
A vector of cluster IDs.
}
\description{
Selects clusters from a set of clusters identified by a clustering method investigate for stratifying signal.
}
\author{
Robert Bruggner
}
