% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.classModel.R,
%   R/citrus.continuousModel.R, R/citrus.model.R
\name{citrus.thresholdCVs}
\alias{citrus.thresholdCVs}
\alias{citrus.thresholdCVs.quick.classification}
\alias{citrus.thresholdCVs}
\alias{citrus.thresholdCVs.quick.continuous}
\alias{citrus.thresholdCVs}
\alias{citrus.thresholdCVs.quick}
\title{Calculate model error rates}
\usage{
citrus.thresholdCVs.quick.classification(modelType, features, labels,
  regularizationThresholds, nCVFolds = 10, ...)

citrus.thresholdCVs.quick.continuous(modelType, features, labels,
  regularizationThresholds, nCVFolds = 10, ...)

citrus.thresholdCVs(modelType, foldFeatures, labels, regularizationThresholds,
  family, folds, foldModels, leftoutFeatures, ...)

citrus.thresholdCVs.quick(modelType, features, labels, regularizationThresholds,
  family, nCVFolds = 10, ...)
}
\arguments{
\item{modelType}{Type of model to be constructed. Valid options are: \code{pamr} and \code{glmnet}.}

\item{features}{Features calculated from a clustering of all samples.}

\item{labels}{Endpoint labels of clustered samples.}

\item{regularizationThresholds}{Thresholds for model regularization.}

\item{nCVFolds}{Number of folds for quick cross-validation.}

\item{...}{Other parameters passsed to model-fitting methods.}

\item{foldFeatures}{List of features with each entry containing features from an independent clustering.}

\item{family}{Model family. Valid options are \code{classification} and \code{continuous}.}

\item{folds}{List of fold indices}

\item{foldModels}{Models constructed from each fold of features.}

\item{leftoutFeatures}{Features calculated for leftout samples mapped to clustered data space.}
}
\value{
Matrix of model error rates, standard error of error estimates, and false discovery rates (if possible) at 
supplied regularization thresholds.
}
\description{
Calculate model error rates at different regularization thresholds.
}
\details{
If independent fold-clustering and fold-features are calculated, use \code{citrus.thresholdCVs}. 
If features are derived from a clustering of all samples together, use \code{citrus.thresholdCVs.quick}. See examples.
}
\author{
Robert Bruggner
}
