% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.util.R
\name{citrus.version}
\alias{citrus.version}
\title{Get Citrus package version}
\usage{
citrus.version()
}
\value{
Citrus package version
}
\description{
Get Citrus package version.
}
\author{
Robert Bruggner
}

