% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_multi_two.R
\docType{data}
\name{d_multitwo}
\alias{d_multitwo}
\title{Combined Boundary Example Dataset}
\format{
A tibble with 5,000 rows and 3 variables:
\describe{
\item{x}{Numeric independent variable generated from a uniform distribution on [-1, 1]}
\item{y}{Numeric independent variable generated from a uniform distribution on [-1, 1]}
\item{class}{Binary categorical variable with two levels: "A" and "B"}
}
}
\usage{
data(d_multitwo)
}
\description{
A demonstration dataset containing two numeric predictors and a binary
class variable. The class boundary is set to
A when x < (-0.5) & y < (-0.3) or
when x >= -0.5 & x < 0.4 & y < 0.1 + 0.8 * x or
when x >= 0.4
B when y < (-1) + 0.8 * x and everywhere else
}
\details{
Simple two-dimensional dataset with a combination of two decision boundaries
}
\keyword{datasets}
