% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumquat.R
\name{generate_perturbations}
\alias{generate_perturbations}
\title{Generate Perturbations Around a Point of Interest}
\usage{
generate_perturbations(
  data,
  poi,
  radius = 0.1,
  step = 0.01,
  predictors = names(poi)
)
}
\arguments{
\item{data}{Training data frame}

\item{poi}{a single row data.frame containing the point of interest}

\item{radius}{Perturbation radius (default: 0.1)}

\item{step}{Step size for perturbations (default: 0.01)}

\item{predictors}{Character vector of predictor variable names}
}
\value{
A data frame of perturbed points. The output contains the following properties:
\itemize{
\item Columns are predictors
\item Number of rows are going to be dependent on \code{radius} and \code{step}
}
}
\description{
Generate Perturbations Around a Point of Interest
}
\examples{
data <- data.frame(
  x = 1,
  y = 2
)
result <- generate_perturbations(
  data,
  poi = data[1,],
  radius = 0.1,
  step = 0.1,
  predictors = c("x", "y")
 )
}
