% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model}
\alias{get_model}
\title{Calculate a prediction model}
\usage{
get_model(features, endpoint, model.type)
}
\arguments{
\item{endpoint}{A vector of length equal to \code{nrow(features)} containing 
the values to be used as prediction endpoints. If this vertor is numeric
a model of family \code{"continuous"} is generated, otherwise if it is
a character or factor vector the model will be of family 
\code{"classification"}}

\item{model.type}{The type of model, either \code{"pamr"}, \code{"sam"} or
\code{"glmnet"}}
}
\value{
A \code{citrus.regressionResult} object with the following properties:
\item{regularizationThresholds}{Regularization thresholds used to constrain all constructed models. Not applicable for \code{sam} models.}
\item{foldModels}{A \code{citrus.endpointModel} constructed from each independent fold feature set. \code{NULL} if \code{nFolds = 1}.}
\item{finalModel}{A \code{citrus.endpointModel} constructed from features derived from the clustering of all samples together.}
\item{thresholdCVRates}{Matrix containing the average error rate and standard error of models at each regularization threshold. FDR also reported where possible.}
\item{cvMinima}{Values and indicies of pre-selected cross-validation error-rate thresholds.}
\item{differentialFeatures}{Non-zero model features and corresponding clusters from the \code{finalModel} constrained by \code{cvMinima}.}
\item{modelType}{Type of model fit on data.}
\item{family}{Family of regression model.}
\item{labels}{Endpoint labels of analyzed samples.}
}
\description{
This function is the main wrapper around the Citrus model building functionality
}
