% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumquat.R
\name{kumquat}
\alias{kumquat}
\title{Complete Local Interpretation Pipeline}
\usage{
kumquat(
  model_bundle,
  data,
  pois,
  perturbations = NULL,
  radius = 0.1,
  step = 0.01,
  predictor_vars = c("x", "y"),
  nfolds = 50,
  alpha = 1,
  class_names = c("A", "B"),
  predict_func = stats::predict
)
}
\arguments{
\item{model_bundle}{A trained model held in a bundle::bundle()}

\item{data}{Training data}

\item{pois}{Points of interest (data rows)}

\item{perturbations}{A list of data.frames of perturbations to be used to fit the local model}

\item{radius}{Perturbation radius (default: 0.1)}

\item{step}{Perturbation step size (default: 0.01)}

\item{predictor_vars}{Character vector of predictor variable names}

\item{nfolds}{Number of CV folds (default: 50)}

\item{alpha}{Elastic net parameter (default: 1)}

\item{class_names}{Character vector of class names}

\item{predict_func}{A function that takes in two arguments: model and data and returns a vector of factors}
}
\value{
A list containing perturbations, predictions, and local model results
}
\description{
Complete Local Interpretation Pipeline
}
\examples{
\dontshow{if (rlang::is_installed(c("randomForest", "bundle"))) withAutoprint(\{ # examplesIf}
data(d_vertical)
rfmodel <- randomForest::randomForest(
 class ~ x + y,
 data = d_vertical
)
# Bundle model up
rfmodel_bundled <- bundle::bundle(rfmodel)
ks <- kumquat(
 rfmodel_bundled,
 d_vertical,
  d_vertical[1,],
  class_names = unique(d_vertical$class)
)
\dontshow{\}) # examplesIf}
}
