% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.plot.R
\name{plot.citrus.regressionResult}
\alias{plot.citrus.regressionResult}
\title{Plot results of a Citrus regression analysis}
\usage{
\method{plot}{citrus.regressionResult}(citrus.regressionResult, outputDirectory,
  citrus.foldClustering, citrus.foldFeatureSet, citrus.combinedFCSSet,
  plotTypes = c("errorRate", "stratifyingFeatures", "stratifyingClusters"),
  byCluster = FALSE, allFeatures = FALSE, ...)
}
\arguments{
\item{citrus.regressionResult}{A \code{citrus.regressionResult} object.}

\item{outputDirectory}{Full path to output directory for plots.}

\item{citrus.foldClustering}{A \code{citrus.foldClustering} object.}

\item{citrus.foldFeatureSet}{A \code{citrus.foldFeatureSet} object.}

\item{citrus.combinedFCSSet}{A \code{citrus.combinedFCSSet} object.}

\item{plotTypes}{Vector of plots types to make. Valid options are \code{errorRate} (Cross-validated error rates for predictive models),
\code{stratifyingFeatures} (plots of non-zero model features),\code{stratifyingClusters} (plots of clustering marker distributions in stratifying clusters)}
}
\description{
Makes many plots showing results of a Citrus analysis
}
\author{
Robert Bruggner
}
