% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_error_rate}
\alias{plot_error_rate}
\title{Plotting model error rate}
\usage{
plot_error_rate(citrus.model, output.file)
}
\arguments{
\item{citrus.model}{An object of class \code{citrus.regressionResult}, as returned from
\code{\link{get_model}}}

\item{output.file}{The path of the output file the plot will be written to}
}
\value{
Returns \code{NULL}
}
\description{
Plotting model error rate
}
