% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_stratifying_features}
\alias{plot_stratifying_features}
\title{Plot stratifying features}
\usage{
plot_stratifying_features(
  citrus.model,
  output.dir,
  by.cluster = FALSE,
  all.features = FALSE
)
}
\arguments{
\item{citrus.model}{An object of class \code{citrus.regressionResult}, as returned from
\code{\link{get_model}}}

\item{output.dir}{The directory that will contain the output. Additional directories and 
sub-directories will be created in \code{output.dir}}

\item{by.cluster}{If this is \code{TRUE} this function will generate a separate plot 
for each cluster, otherwise or a single plot with all the clusters. 
The latter option may take quite long to plot if there are several clusters}

\item{all.features}{Whether to plot only the stratifying features (default) or all the features}
}
\description{
Plot stratifying features
}
