% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{run_analysis}
\alias{run_analysis}
\title{High level wrapper}
\usage{
run_analysis(citrus.features, endpoint, output.directory, model.type,
  plot.by.cluster = FALSE, plot.all.features = FALSE,
  clusters.data = NULL)
}
\description{
High level wrapper
}
