% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citrus.plot.R
\name{citrus.plotClusters}
\alias{citrus.plotClusters}
\title{Plot cluster histograms}
\usage{
citrus.plotClusters(clusterIds, clustersData, conditions = NULL,
  outputDir = NULL)
}
\arguments{
\item{clusterIds}{Vector of cluster IDs to plot}

\item{conditions}{Vector of conditions clustering was performed on.}

\item{clusterAssignments}{List containing indicies of cells assigned to each cluster.}

\item{citrus.combinedFCSSet}{Combined FCS data that was clustered.}

\item{clusteringColumns}{Columns for which to plot distributions}

\item{outputFile}{If not \code{NULL}, plot is written to \code{outputFile}.}
}
\description{
Plot expression of markers in cluster cells relative to all cells
}
\author{
Robert Bruggner
}
