% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumquat.R
\name{generate_perturbations}
\alias{generate_perturbations}
\title{Generate Perturbations Around a Point of Interest}
\usage{
generate_perturbations(
  data,
  poi,
  radius = 0.1,
  step = 0.01,
  x_var = "x",
  y_var = "y"
)
}
\arguments{
\item{data}{Training data frame}

\item{poi}{Row number of point of interest}

\item{radius}{Perturbation radius (default: 0.1)}

\item{step}{Step size for perturbations (default: 0.01)}

\item{x_var}{Name of x variable (default: "x")}

\item{y_var}{Name of y variable (default: "y")}
}
\value{
A data frame of perturbed points
}
\description{
Generate Perturbations Around a Point of Interest
}
