% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kumquat.R
\name{kumquat}
\alias{kumquat}
\title{Complete Local Interpretation Pipeline}
\usage{
kumquat(
  model_bundle,
  data,
  pois,
  perturbations = NULL,
  radius = 0.1,
  step = 0.01,
  predictor_vars = c("x", "y"),
  nfolds = 50,
  alpha = 1,
  class_names = c("A", "B")
)
}
\arguments{
\item{model_bundle}{A trained model held in a bundle::bundle()}

\item{data}{Training data}

\item{pois}{Points of interest (row numbers)}

\item{perturbations}{A data.frame of perturbations to be used to fit the local model}

\item{radius}{Perturbation radius (default: 0.1)}

\item{step}{Perturbation step size (default: 0.01)}

\item{predictor_vars}{Character vector of predictor variable names}

\item{nfolds}{Number of CV folds (default: 50)}

\item{alpha}{Elastic net parameter (default: 1)}

\item{class_names}{Character vector of class names}
}
\value{
A list containing perturbations, predictions, and local model results
}
\description{
Complete Local Interpretation Pipeline
}
