% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pinch_importance}
\alias{pinch_importance}
\title{Extract and Combine Feature Importances from Kumquat Objects}
\usage{
pinch_importance(kumquats)
}
\arguments{
\item{kumquats}{A result from a call to \code{make_kumquats}}
}
\value{
A \code{data.frame} where each row corresponds to the feature
importances extracted from one kumquat object. Column names represent
feature names.
}
\description{
\code{pinch_importance()} extracts feature importance values from
kumquats and combines them into a single data frame.
}
\examples{
\dontrun{
pinch_importance(make_kumquats(model_bundle, dataset, pois, class_names))
}


}
